Comments and clean up FT UI
