+++ packages/front-end/components/FactTables/FactTableModal.tsx
@@ -8,6 +8,8 @@ import { useRouter } from "next/router";
 import { isProjectListValidForProject } from "shared/util";
 import { useEffect, useState } from "react";
 import { FaExternalLinkAlt } from "react-icons/fa";
+import Collapsible from "react-collapsible";
+import { PiCaretRightFill } from "react-icons/pi";
 import { useDefinitions } from "@/services/DefinitionsContext";
 import { useAuth } from "@/services/auth";
 import useOrgSettings from "@/hooks/useOrgSettings";
@@ -20,6 +22,9 @@ import { getNewExperimentDatasourceDefaults } from "@/components/Experiment/NewE
 import Code from "@/components/SyntaxHighlighting/Code";
 import { usesEventName } from "@/components/Metrics/MetricForm";
 import EditFactTableSQLModal from "@/components/FactTables/EditFactTableSQLModal";
+import AggregatedFactTableSettings, {
+  getAggregatedFactTableSettingsFormDefault,
+} from "@/components/FactTables/AggregatedFactTableSettings";
 import { useUser } from "@/services/UserContext";
 import Checkbox from "@/ui/Checkbox";
 import { getAutoSliceUpdateFrequencyHours } from "@/services/env";
@@ -70,10 +75,16 @@ export default function FactTableModal({
       managedBy: existing?.managedBy || "",
       projects: existing?.projects || [],
       autoSliceUpdatesEnabled: existing?.autoSliceUpdatesEnabled ?? false,
+      aggregatedFactTableSettings:
+        getAggregatedFactTableSettingsFormDefault(existing),
     },
   });
 
   const selectedDataSource = getDatasourceById(form.watch("datasource"));
+
+  const datasourceHasIncrementalRefresh =
+    selectedDataSource?.settings?.pipelineSettings?.allowWriting === true &&
+    selectedDataSource?.settings?.pipelineSettings?.mode === "incremental";
 
   useEffect(() => {
     if (!selectedDataSource || existing) return;
@@ -141,6 +152,20 @@ export default function FactTableModal({
           // Default eventName to the metric name
           value.eventName = value.eventName || value.name;
 
+          // Clearing all id types disables the aggregated pipeline.
+          const aggSettings = value.aggregatedFactTableSettings;
+          const aggIdTypes = (aggSettings?.idTypes ?? []).filter((id) =>
+            value.userIdTypes.includes(id),
+          );
+          const normalizedAggSettings =
+            aggIdTypes.length && aggSettings
+              ? {
+                  idTypes: aggIdTypes,
+                  updateTime: aggSettings.updateTime,
+                  lookbackWindow: Number(aggSettings.lookbackWindow),
+                }
+              : null;
+
           if (existing && !duplicate) {
             const data: UpdateFactTableProps = {
               description: value.description,
@@ -152,6 +177,9 @@ export default function FactTableModal({
               projects: value.projects,
               autoSliceUpdatesEnabled: value.autoSliceUpdatesEnabled,
             };
+            if (hasCommercialFeature("pipeline-mode")) {
+              data.aggregatedFactTableSettings = normalizedAggSettings;
+            }
             await apiCall(`/fact-tables/${existing.id}`, {
               method: "PUT",
               body: JSON.stringify(data),
@@ -182,6 +210,8 @@ export default function FactTableModal({
             }
             value.columns = [];
             value.projects = projects;
+            value.aggregatedFactTableSettings =
+              normalizedAggSettings ?? undefined;
 
             const { factTable, error } = await apiCall<{
               factTable: FactTableInterface;
@@ -300,6 +330,32 @@ export default function FactTableModal({
             />
           </div>
         )}
+
+        {!!existing &&
+          hasCommercialFeature("pipeline-mode") &&
+          datasourceHasIncrementalRefresh && (
+            <>
+              <hr className="mt-4" />
+              <Collapsible
+                trigger={
+                  <div className="link-purple font-weight-bold mt-2 mb-2">
+                    <PiCaretRightFill className="chevron mr-1" />
+                    Advanced Settings
+                  </div>
+                }
+                transitionTime={100}
+                lazyRender={true}
+              >
+                <div className="rounded px-3 pt-3 pb-1 bg-highlight">
+                  <AggregatedFactTableSettings
+                    form={form}
+                    userIdTypes={form.watch("userIdTypes")}
+                    canEdit={permissionsUtil.canUpdateFactTable(existing, {})}
+                  />
+                </div>
+              </Collapsible>
+            </>
+          )}
       </Modal>
     </>
   );
