+++ packages/front-end/pages/fact-tables/[ftid].tsx
@@ -17,7 +17,6 @@ import FactTableModal from "@/components/FactTables/FactTableModal";
 import Code from "@/components/SyntaxHighlighting/Code";
 import ColumnList from "@/components/FactTables/ColumnList";
 import AggregatedFactTablesCard from "@/components/FactTables/AggregatedFactTablesCard";
-import AggregatedFactTableSettings from "@/components/FactTables/AggregatedFactTableSettings";
 import FactFilterList from "@/components/FactTables/FactFilterList";
 import EditProjectsForm from "@/components/Projects/EditProjectsForm";
 import PageHead from "@/components/Layout/PageHead";
@@ -560,11 +559,6 @@ export default function FactTablePage() {
           </TabsContent>
         </Box>
       </Tabs>
-
-      <AggregatedFactTableSettings
-        factTable={factTable}
-        mutate={mutateDefinitions}
-      />
 
       <AggregatedFactTablesCard factTable={factTable} />
     </div>
